[chore] HashtagChipCell 파일 위치 조정
